refactor(app-bar): move to folder structure and use sign-in-button
- Move app-bar.tsx into app-bar/ folder
- Reduce title size on mobile (text-2xl) for better fit
- Use new SignInButton component with GitHub auth support
